SL.Final.Rmd
Now updated with LDA and QDA in the models. QDA was finnicky. It had trouble with some "rank-deficient" data in the various training sets, but I customized the code to exclude the two variables that kept giving us issues due to that: social.smoker and Disciplinary.failure. Once those were out of the QDA analysis, everything else ran fine.
With that, I think we have everything we possibly need in our code!